refactor(node): thin allocdb-jepsen binary
